Rename "*.hpp" to "*.h" for accesor tests
